% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{aPb_series_out}
\alias{aPb_series_out}
\title{Calculate aPb, aPPb, aPPPPb, tr(P), tr(PP) (and P)}
\usage{
aPb_series_out(a, b, u, ev, x)
}
\description{
Calculate aPb, aPPb, aPPPPb, tr(P), tr(PP) (and P)
}
