% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EMM2_last_step}
\alias{EMM2_last_step}
\title{Calculate Vu, Ve, LL and Hinv when given lambda}
\usage{
EMM2_last_step(lambda, y, x, u, d, z, k, th, ome_sq, ph, maxval, n, p, df,
  SE = FALSE, return_Hinv = FALSE)
}
\description{
Calculate Vu, Ve, LL and Hinv when given lambda
}
