% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GWAS_F_test}
\alias{GWAS_F_test}
\title{Calculate statistic of GWAS which follows Beta distribution}
\usage{
GWAS_F_test(y, x, hinv, v1, v2, p)
}
\description{
Calculate statistic of GWAS which follows Beta distribution
}
