% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hess_func_out}
\alias{hess_func_out}
\title{Calculate the score function}
\usage{
hess_func_out(n, p, tr_HinvG2, yPy, yPGPy, yPGPGPy, tr_PG2, REML = TRUE)
}
\description{
Calculate the score function
}
