% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch the rainbow Shiny app}
\usage{
launch_app()
}
\value{
Starts the Shiny application
}
\description{
This function launches the interactive Shiny app included in the package.
The app allows users to explore nitrate concentration by site.
}
\examples{
# rainbow::launch_app()
}
