% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makerainbows.R
\name{makerainbows}
\alias{makerainbows}
\title{Make rainbows}
\usage{
makerainbows(...)
}
\arguments{
\item{...}{Vector specifying the boxes for each band}
}
\value{
Dataframe
}
\description{
Create dataframe for rainbow plot
}
\examples{
df<-rainbowdf(1,1,1,1,1,1,1,1);
df
}
