% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ml_est_out}
\alias{ml_est_out}
\title{Maximize likelihood when given lambda_0}
\usage{
ml_est_out(lambda, y, x, u, delta, z, k, logXtX, conv_param = 1e-06,
  count_max = 15L, bounds1 = 1e-06, bounds2 = 1e+06, REML = TRUE)
}
\description{
Maximize likelihood when given lambda_0
}
