% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ml_one_step}
\alias{ml_one_step}
\title{Calculate some values when given lambda}
\usage{
ml_one_step(lambda, y, x, u, delta, z, k, logXtX, conv_param = 1e-06,
  count_max = 15L, bounds1 = 1e-06, bounds2 = 1e+06, REML = TRUE,
  SE = FALSE, return_Hinv = TRUE)
}
\description{
Calculate some values when given lambda
}
