% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{my_first_plot}
\alias{my_first_plot}
\title{My First Plot
Plots a simple \code{ggplot2} boxplot of the \code{rivers_data} dataset,
with Site on the x-axis and Nitrate on the y-axis.}
\usage{
my_first_plot()
}
\value{
A \code{ggplot2} object containing the plot.
}
\description{
This function visualises nitrate concentration differences between
sampling sites in the synthetic \code{rivers_data} dataset. It uses a clean minimal theme and returns the ggplot object so it can be
further customised.
}
\examples{
devtools::load_all()
my_first_plot()
}
