% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{P_calc}
\alias{P_calc}
\title{Calculate P or Hinv, and log(P) or log(Hinv)}
\usage{
P_calc(lambda, Ws, Gammas, u, d, lw, Pornot = TRUE, gammas_diag = TRUE)
}
\description{
Calculate P or Hinv, and log(P) or log(Hinv)
}
