% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotrainbows.R
\name{plotrainbows}
\alias{plotrainbows}
\title{Plot rainbows}
\usage{
plotrainbows(
  df,
  shape,
  colourby = "individual",
  colourpal = "",
  alpha = 0.9,
  labels,
  innercircletext = ""
)
}
\arguments{
\item{df}{Dataframe from makerainbows function}

\item{shape}{Specify "circle" or "semi-circle"}

\item{colourby}{Colours determined by individual box - "individual" or by band - "band"; default is individual box}

\item{colourpal}{Colour palette - can specify "GilbertBaker1978"}

\item{alpha}{Vector of alpha values for boxes, specify centre ring clockwise outward; default is 0.9}

\item{labels}{Vector of labels for boxes, specify centre ring clockwise outward}

\item{innercircletext}{Vector of labels for boxes, specify centre ring clockwise outward; default is blank}
}
\value{
ggplot2 rainbow plot
}
\description{
Create a rainbow plot using ggplot2
}
\examples{

# Wider Determinants of Health
df <- makerainbows(1,1,6,1)
labelsWDH<-c("Individual Lifestyle Factors",
             "Social and Community Networks",
             "Food prod.", "Education", "Work Env.", "Water/Sanitation", "Healthcare", "Housing",
             "General Socioeconomic, Cultural and Environmental Factors")
plotrainbows(df, shape="semi-circle", colourby="band", colourpal="",labels=labelsWDH, innercircletext = "Age, sex and constitutional factors")

Place Standard


Doughnut (economic model)


# Pride Rainbow
df <- makerainbows(1,1,1,1,1,1,1,1)
plotrainbows(df, shape="semi-circle", colourby="individual", colourpal="GilbertBaker1978", labels=c("Sprit", "Serenity", "Magic", "Nature", "Sunlight", "Healing","Life", "Sex"));
}
