% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbExtractDaily.r
\name{rbExtractDaily}
\alias{rbExtractDaily}
\alias{rbExtractMonthly}
\alias{rbExtractYearly}
\title{Extract daily values of weather variables from a PRISM raster data set}
\usage{
rbExtractDaily(
  x,
  prismDir,
  yearField = "year",
  monthField = "month",
  domField = "day",
  dateField = NULL,
  longLat = NULL,
  prismSet = "an81",
  rastVars = c("tmin", "tmax", "ppt"),
  rastSuffix = "bil",
  windowYears = 0,
  windowDays = 7,
  verbose = TRUE
)

rbExtractMonthly(
  prismDir,
  prismSet = "lt81",
  rastVars,
  x,
  yearField = "year",
  monthField = "month",
  domField = "day",
  dateField = NULL,
  longLat = NULL,
  windowYears = 0,
  windowMonths = 0,
  verbose = TRUE
)

rbExtractYearly(
  prismDir,
  prismSet = "lt81",
  rastVars,
  x,
  yearField = "year",
  monthField = "month",
  domField = "day",
  dateField = NULL,
  longLat = NULL,
  windowYears = 0,
  verbose = TRUE
)
}
\arguments{
\item{x}{Data frame.}

\item{prismDir}{Character. Path of the base directory in which the rasters are stored. Folder contains a subfolder named \code{an81} or \code{lt81}.}

\item{yearField, monthField, domField}{Character. Names of columns in \code{x} with year, month, and day of month of observation. These must all be integers or \code{NA} values. You can specify these values or use argument \code{dateField} to specify a column with dates.}

\item{dateField}{Name of column in \code{x} with date of each record. Values must be of in YYYY-MM-DD (year-month-day of month) format \emph{or} already of class \code{\link{Date}}. See \code{\link[lubridate]{ymd}} or related functions for more help.}

\item{longLat}{Character vector with two elements. Names of columns in \code{x} with longitude and latitude (in that order). Coordinates will be assumed to be in the same coordinate reference system as the PRISM rasters.}

\item{prismSet}{Character. Name of PRISM data subset. This can be either \code{an81} (daily weather data) or \code{lt81} (monthly and annual weather data).}

\item{rastVars}{Name of variable(s) to extract. Valid values are:
\itemize{
    \item    \code{ppt}: Precipitation.
    \item    \code{tmax}: Maximum temperature.
    \item    \code{tmin}: Minimum temperature.
    \item    \code{tmean}: Mean temperature.
    \item    \code{tdmean}: Dew-point temperature.
    \item    \code{vpdmax}: Maximum vapor-pressure.
    \item    \code{vpdmin}: Minimum vapor-pressure.
}}

\item{rastSuffix}{Character. The "suffix" at the end of each weather raster. PRISM rasters are usually shipped in 'BIL' format, so normally this should be \code{bil}. However, any other suffix corresponding to a raster type that can be opened by the \code{\link[terra]{rast}} function can be used if the rasters have been converted to another format.}

\item{windowYears, windowMonths, windowDays}{Integers >= 0. Extract data for this many years, months, and/or days before the day of each observation, \emph{plus} the day of observation. Note:
\itemize{
    \item If \code{prismSet} is \code{an81} (daily), then only \code{windowYears} and \code{windowDays} are used. Note that the difference between using, say, \code{windowYears = 1} and \code{windowDays = 365} is that the former can accommodate leap days whereas the latter just extracts the 365 days prior (which may be less than a full calendar year if the timespan encompasses a leap day).
    \item If \code{prismSet} is \code{lt81} (monthly), then only \code{windowYears} and \code{windowMonths} are used.
    \item If \code{prismSet} is \code{lt81} (annual), then only \code{windowYears}is used.
}
To get only data for the day/month/year of each record, set all of the respective \code{window} arguments to 0.}

\item{verbose}{Logical. If \code{TRUE} (default), show progress.}
}
\value{
Matrix with one row per row in \code{x}. \code{NA} values represent days/months/years that did not fall within the specified window or for which rasters were unavailable.
}
\description{
These functions extract values from interpolated weather rasters from the Parameter Regression on Independent Slopes (PRISM) data product (Daly et al. 2002). Depending on the set, PRISM rasters represent interpolated values of daily, monthly, or annual weather. Extractions are done at points (versus polygons or lines, for example). \cr\cr
The basic input is data frame or data frame-like object, with each row representing a location and having a particular date. The function also needs to be pointed toward a folder with PRISM data. The function assumes that the rasters are arranged in the manner in which they are normally supplied by the PRISM Climate Group. That is, for example, each daily/monthly/annual raster is inside a folder corresponding to the year, which in turn is in a folder named either \code{daily} or \code{monthly}, which is inside a folder named after the variable the rasters represent (e.g., \code{tmin}, \code{vpdmax}, etc.), which is inside a folder named either \code{an81} (daily values) or \code{lt81} (monthly and annual values). Not all years or variables are necessarily available, as a license must be purchased and data acquired from the PRISM Climate Group. \cr\cr
The function can extract values corresponding to the day/month/year of each record, plus (optionally) a user-specified window of time prior to the day/month/year of each record. For example, you could use this to extract daily climate data for a site collected on April 22, 2014, and all days prior up to 10 years (April 23, 2004). This function is really a fancy wrapper for \code{\link[terra]{extract}}, but it does the tough job of getting the directory structures right, pulling all needed rasters, and efficiently grouping records to speed extraction. \cr\cr
The function does not assume that data for all PRISM years are available (i.e., if you only have rights to the data from, say, 2015 to 2018), but it does assume that all relevant rasters for a particular year are available within each yearly folder. If rasters preceding a date only partially cover the window, then values for the part covered will be extracted. For example if you try to extract annual values for a window spanning 2010 to 2020 but only have available rasters for 1981 to 2018, then values for 2010 to 2018 will be extracted.
}
\section{Functions}{
\itemize{
\item \code{rbExtractMonthly}: Extract monthly values of weather variables from a PRISM raster data set

\item \code{rbExtractYearly}: Extract annual values of weather variables from a PRISM raster data set
}}

\examples{
\dontrun{
x <- data.frame(
	long=rep(-97.66, 5),
	lat=rep(38.37, 5),
	date=c('2015-12-31', '1981-01-05', '2020-12-01',
	'2019-01-05', '1895-05-01')
)

y <- rbExtractDaily(
	prismDir = 'F:/ecology/Climate/PRISM acquired in 2020',
	prismSet = 'an81',
	rastVars = 'tmin',
 rastSuffix = 'tif',
	x = x,
	yearField = NULL,
	monthField = NULL,
	domField = NULL,
	dateField = 'date',
	longLat = c('long', 'lat'),
	windowYears = 0,
	windowDays = 7,
	verbose = TRUE
)

y <- rbExtractMonthly(
	prismDir = 'F:/ecology/Climate/PRISM acquired in 2020',
	prismSet = 'lt81',
	rastVars = 'tmin',
 rastSuffix = 'tif',
	x = x,
	yearField = NULL,
	monthField = NULL,
	domField = NULL,
	dateField = 'date',
	longLat = c('long', 'lat'),
	windowYears = 1,
	windowMonths = 1,
	verbose = TRUE
)

y <- rbExtractYearly(
	prismDir = 'F:/ecology/Climate/PRISM acquired in 2020',
	prismSet = 'lt81',
	rastVars = 'tmin',
 rastSuffix = 'tif',
	x = x,
	yearField = NULL,
	monthField = NULL,
	domField = NULL,
	dateField = 'date',
	longLat = c('long', 'lat'),
	windowYears = 3,
	verbose = TRUE
)

}
}
\references{
Daly, C., Gibson, W.P., Taylor, G.H., Johnson, G.L., and Pasteris, P.  2002.  A knowledge-based approach to the statistical mapping of climate. \emph{Climate Research} 22:99-113. \href{http://dx.doi.org/10.3354/cr022099}{DOI: 10.3354/cr022099}
}
