% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RGWAS.menu.R
\name{RGWAS.menu}
\alias{RGWAS.menu}
\title{Print the R code which you should perform for RAINBOW GWAS}
\usage{
RGWAS.menu()
}
\value{
the R code which you should perform for RAINBOW GWAS
}
\description{
Print the R code which you should perform for RAINBOW GWAS
}
