% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbdata.R
\docType{data}
\name{sbdata}
\alias{sbdata}
\title{Climate Data from Santa Barbara}
\format{A data frame with 3287 rows and 11 colums
\describe{
\item{date}{Year-Month-Day}
\item{day}{(1-31)}
\item{month}{(1-12)}
\item{year}{ (integer)}
\item{wy}{year starting Oct 1 (integer) }
\item{wyd}{day of water year  (integer) }
\item{tavg}{air temperature (degrees C) }
\item{vpg}{vapor pressure deficit (MPa) }
\item{rnet}{net radiation (kJ/day) }
\item{preip}{precipitation (mm/day) }
}}
\source{
\url{http://sbc.lternet.edu/data/}
}
\usage{
sbdata
}
\description{
Data from SB-LTER met station, daily meterology
}
\keyword{datasets}
