% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{score_fisher}
\alias{score_fisher}
\title{Calculate fisher information of the log-likelihood for score test}
\usage{
score_fisher(p0, Gs_all, nuisance_no, lg_all)
}
\description{
Calculate fisher information of the log-likelihood for score test
}
