% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{score_fisher_linker_diag}
\alias{score_fisher_linker_diag}
\title{Calculate fisher information of the log-likelihood for score test (for linear kernel)}
\usage{
score_fisher_linker_diag(p0, Gs_all, nuisance_no, lw_all)
}
\description{
Calculate fisher information of the log-likelihood for score test (for linear kernel)
}
