% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{score_l1}
\alias{score_l1}
\title{Calculate first derivariate of the log-likelihood for score test}
\usage{
score_l1(y, p0, Gs, lg)
}
\description{
Calculate first derivariate of the log-likelihood for score test
}
