% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{score_l1_linker}
\alias{score_l1_linker}
\title{Calculate first derivariate of the log-likelihood for score test (for linear kernel)}
\usage{
score_l1_linker(y, p0, Ws, Gammas, lw)
}
\description{
Calculate first derivariate of the log-likelihood for score test (for linear kernel)
}
