% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{score_l1_linker_diag}
\alias{score_l1_linker_diag}
\title{Calculate first derivariate of the log-likelihood for score test (for linear kernel)}
\usage{
score_l1_linker_diag(y, p0, W2s, lw)
}
\description{
Calculate first derivariate of the log-likelihood for score test (for linear kernel)
}
