% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{spectralG_cholesky}
\alias{spectralG_cholesky}
\title{The spectral decomposition of G matrix (cholesky decomposition)}
\usage{
spectralG_cholesky(zbt, x, return_G = TRUE, return_SGS = FALSE)
}
\description{
The spectral decomposition of G matrix (cholesky decomposition)
}
