% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{spectralG_eigen}
\alias{spectralG_eigen}
\title{The spectral decomposition of G matrix (eigen decomposition)}
\usage{
spectralG_eigen(zkzt, x, return_G = TRUE, return_SGS = FALSE)
}
\description{
The spectral decomposition of G matrix (eigen decomposition)
}
