% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spring_summary.R
\name{spring_summary}
\alias{spring_summary}
\title{Summary information about spring climate}
\usage{
spring_summary(clim_data, spring_months = c(4:6), spring_out = FALSE)
}
\arguments{
\item{clim_data}{data frame with columns tavg (C)
rain (precip in mm), year, month (integer), day}

\item{spring_out}{(default FALSE) set to TRUE to output spring precip and temperature for all years}

\item{months}{(as integer) to include in spring; default 4,5,6}
}
\value{
returns a list containing,
\describe{
 \item{mean_springT}{mean_springT mean spring temperature (C)}
 \item{coldest_spring_yr}{ year with lowest spring temperature  (year)}
 \item{warmest_spring_yr}{ year with highest spring temperature  (year)}
 \item{coldest_springT}{  lowest spring temperature  (C)}
 \item{warmest_springT}{  highest spring temperature  (C)}
 \item{mean_springP}{ mean spring precipitation  (mm)}
 \item{wettest_spring_yr}{ spring (as year) with highest precip (year)}
 \item{wettest_springP}{  highest precip (mm)}
 \item{driest_spring_yr}{ spring (as year) with highest precip (year)}
 \item{driest_springP}{  highest precip (mm)}
 and array of means for all years if spring_out set to TRUE
 }
}
\description{
computes summary information about spring temperature and precipitation
}
