\name{rainbow-package}
\alias{rainbow-package}
\alias{rainbow}
\docType{package}
\title{
Rainbow plot, functional bagplot, functional HDR boxplot
}
\description{
This package computes the rainbow plot, functional bagplot, functional HDR boxplot. The latter two can also be used to identify outliers, which have either the lowest depth or the lowest density.
}
\author{
Han Lin Shang and Rob J Hyndman

Maintainer: Han Lin Shang <HanLin.Shang@buseco.monash.edu.au>
}
\references{
R. J. Hyndman and H. L. Shang. (2008) "Rainbow plots, bagplots, and boxplots for functional data", \emph{Tech. rep. Department of Econometrics \& Business Statistics}, Monash University, November 2008, \url{http://www.buseco.monash.edu.au/depts/ebs/pubs/wpapers/2008/9-08.php}.
}
\keyword{package}

