\name{rainbow-package}
\alias{rainbow-package}
\alias{rainbow}
\docType{package}
\title{
Rainbow plots, bagplots and boxplots for functional data
}
\description{
This package computes the rainbow plots, bagplots and boxplots for functional data. The latter two can also be used to identify outliers, which have either the lowest depth or the lowest density.
}
\author{
Han Lin Shang and Rob J Hyndman

Maintainer: Han Lin Shang <HanLin.Shang@buseco.monash.edu.au>
}
\references{
R. J. Hyndman and H. L. Shang. (2009) "Rainbow plots, bagplots, and boxplots for functional data", \emph{Journal of Computational and Graphical Statistics}, \bold{in press}.
}
\keyword{package}

