\name{rainbow-package}
\alias{rainbow-package}
\alias{rainbow}
\docType{package}
\title{
Rainbow Plots, Bagplots and Boxplots for Functional Data
}
\description{
This package presents the rainbow plots, bagplots and boxplots for functional data. The latter two can also be used to identify outliers, which have either the lowest depth or the lowest density, respectively. 
}
\author{
Han Lin Shang and Rob J Hyndman

Maintainer: Han Lin Shang <hanlin.shang@anu.edu.au>
}
\references{
R. J. Hyndman and H. L. Shang. (2008) "Bagplots, boxplots and outlier detection for functional data", in S. Dabo-Niang and F. Ferraty, eds, `Functional and Operatorial Statistics', Springer, Heidelberg, pp. 201-207.

R. J. Hyndman and H. L. Shang. (2010) "Rainbow plots, bagplots, and boxplots for functional data", \emph{Journal of Computational and Graphical Statistics}, \bold{19}(1), 29-45.

H. L. Shang (2011) "rainbow: an R package for visualizing functional time series", \emph{The R Journal}, \bold{3}(2), 54-59.

H. L. Shang and R. J. Hyndman (2011) "Nonparametric time series forecasting with dynamic updating", Mathematics and Computers in Simulation, \bold{81}(7), 1310-1324.

H. L. Shang (2013) "Functional time series approach for forecasting very short-term electricity demand", Journal of Applied Statistics, \bold{40}(1), 153-168.

H. L. Shang (2013) "ftsa: An R package for analyzing functional time series", The R Journal, \bold{5}(1), 64-72.

H. L. Shang (2014) "A survey of functional principal component analysis", Advances in Statistical Analysis, \bold{98}(2), 121-142.

H. L. Shang, P. Smith, J. Bijak and A. Wisniowski (2016) "A multilevel functional data method for forecasting population, with an application to the United Kingdom", International Journal of Forecasting, \bold{32}(3), 629-649.

H. L. Shang (2016) "Mortality and life expectancy forecasting for a group of populations in developed countries: A multilevel functional data method", Annals of Applied Statistics, \bold{10}(3), 1639-1672.

H. L. Shang (2017) "Functional time series forecasting with dynamic updating: An application to intraday particulate matter concentration", Econometrics and Statistics, 1, 184-200.

H. L. Shang (2017) "Forecasting intraday S&P 500 index returns: A functional time series approach", Journal of Forecasting, \bold{36}(7), 741-755.

H. L. Shang and S. Haberman (2017) "Grouped multivariate and functional time series forecasting: An application to annuity pricing", Insurance: Mathematics and Economics, \bold{75}, 166-179.

H. L. Shang and R. J. Hyndman (2017) "Grouped functional time series forecasting: An application to age-specific mortality rates", Journal of Computational and Graphical Statistics, \bold{26}(2), 330-343.

G. Rice and H. L. Shang (2017) "A plug-in bandwidth selection procedure for long-run covariance estimation with stationary functional time series", Journal of Time Series Analysis, 38(4), 591-609.

P. Reiss, J. Goldsmith, H. L. Shang and R. T. Ogden (2017) "Methods for scalar-on-function regression", International Statistical Review, \bold{85}(2), 228-249.

P. Kokoszka, G. Rice and H. L. Shang (2017) "Inference for the autocovariance of a functional time series under conditional heteroscedasticity", Journal of Multivariate Analysis, \bold{162}, 32-50.

Y. Gao and H. L. Shang (2017) "Multivariate functional time series forecasting: An application to age-specific mortality rates", Risks, \bold{5}(2), 21.

H. L. Shang (2018) "Visualizing rate of change: An application to age-specific fertility rates", Journal of the Royal Statistical Society: Series A (Statistics in Society), \bold{182}(1), 249-262.

H. L. Shang (2018) "Bootstrap methods for stationary functional time series", Statistics and Computing, \bold{28}(1), 1-10/

Y. Gao, H. L. Shang and Y. Yang (2018) "High-dimensional functional time series forecasting: An application to age-specific mortality rates", Journal of Multivariate Analysis, \bold{forthcoming}.
}
\keyword{package}

