% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAINBOW.R
\docType{package}
\name{RAINBOW}
\alias{RAINBOW}
\alias{RAINBOW-package}
\title{RAINBOW :
Perform genome wide asscoiation study (GWAS) by kernel-based methods}
\description{
Users can test multiple SNPs simultaneously by kernel-based methods. Users can test not only additive effects but also dominance and epistatic effects.
}
